requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
yfinance
scipy
pillow